requirement.txt
Django
python-dotenv
django-crispy-forms
crispy-bootstrap5
stripe
gunicorn
whitenoise
Pillow